+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## v1.1.1
+### Changed
+ - Addon compartment improved
+
+## v1.1.0
+### Changed
+ - Addon compartment added
+
 ## v1.0.17
 ### Changed
  - Updated ToC version